moving more files to specific package
