+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/BanksDropdownMenu.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.components
+package br.com.rbrthmn.ui.financialcompanion.screens.home.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.shape.CircleShape
