+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/CreditCardBillsCard.kt
@@ -57,10 +57,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
-import br.com.rbrthmn.ui.financialcompanion.components.AddItemButton
-import br.com.rbrthmn.ui.financialcompanion.components.BanksDropdownMenu
 import br.com.rbrthmn.ui.financialcompanion.components.DecimalInputField
-import br.com.rbrthmn.ui.financialcompanion.components.TotalValueText
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
