+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/operations/components/DatePickerDocked.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.components
+package br.com.rbrthmn.ui.financialcompanion.screens.operations.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
